(web) add static website for displaying results tables
Create a static web site using Python and Jinja as the template
processor (instead of a hodgepodge M4 and Perl).

- now have a working `make serve` with an index

- move old M4/Perl-based stuff into "the attic"

- move `samples.xslx` to `supplemental_data` dir

- add `devsetup` make target

- create a Python decorator to load TOML configs and, e.g., make those
  available for the template processor

Historical note: the original web site prototype was based on earlier
work for the vTR project, which used M4 (primarily) and the
Template::Toolkit Perl library (for its looping constructs) to generate
the static site.

This was later replaced with Python code using Jinja.

The `oldweb` branch in the original GitLab repository preserves the
original commits, if they're of archaeological interest later.